Lazily invert FrozenMappings only when it's actually needed
This seemed like an obvious inefficiency, although I don't have any
profiling information which says it's strictly nessicarry.
Unfortunately this makes inverting a mapping slightly slower,
since there is a little bit of overhead for the laziness
and the space-efficient representation.
The overhead should be negligable.

One extra benefit is that a FrozenMapping now contains
only one Arc instead of two.
It's still the same size (due to extra info), but cloning only requires
incrementing one Arc instead of two,
This is (in theory) a relatively expensive operation.