Move all the map_class functions into a MapClass trait
This is all based around a new TypeTransformer,
which makes it
This also opens up the possibility for caching remapped signatures,
since I've addeed a remap_signature function to TypeTransformer.
Right now none of the mappings, since FrozenMappings must be sync and
I'd like to avoid locking or thread local data.